requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
Django==5.1.3
django-colorfield==0.11.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
markupsafe==3.0.2
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.45.1
whitenoise==6.8.2
